requirements.txt
aiohttp==3.8.3
aiohttp-socks==0.7.1
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
exceptiongroup==1.0.0rc9
fake-useragent==0.1.11
frozenlist==1.3.1
h11==0.14.0
idna==3.4
multidict==6.0.2
outcome==1.2.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-socks==2.0.3
pytz==2022.4
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
urllib3==1.26.12
webencodings==0.5.1
wsproto==1.2.0
yarl==1.8.1
img2pdf~=0.4.4
beautifulsoup4~=4.11.1
asyncio~=3.4.3
aiofiles~=22.1.0
selenium~=4.5.0
aiogram~=2.23.1